requirements.txt
streamlit
pandas
numpy
requests
python-dateutil
python-dotenv
